requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==1.10.12
google-generativeai==0.3.2
PyPDF2==3.0.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2